thunkgen: Passthrough abi to analysis
