+++ ThunkLibs/Generator/analysis.h
@@ -8,7 +8,88 @@
 #include <string>
 #include <unordered_map>
 #include <unordered_set>
+#include <variant>
 #include <vector>
+
+struct SimpleTypeInfo {
+  uint64_t size_bits;
+  uint64_t alignment_bits;
+
+  bool operator==(const SimpleTypeInfo& other) const {
+    return size_bits == other.size_bits && alignment_bits == other.alignment_bits;
+  }
+};
+
+struct StructInfo : SimpleTypeInfo {
+  struct MemberInfo {
+    uint64_t size_bits; // size of this member. For arrays, total size of all elements
+    uint64_t offset_bits;
+    std::string type_name;
+    std::string member_name;
+    std::optional<uint64_t> array_size;
+    bool is_function_pointer;
+    bool is_integral;
+    bool is_signed_integer;
+
+    bool operator==(const MemberInfo& other) const {
+      return size_bits == other.size_bits && offset_bits == other.offset_bits &&
+             // The type name may differ for integral types if all other parameters are equal
+             (type_name == other.type_name || (is_integral && other.is_integral)) && member_name == other.member_name &&
+             array_size == other.array_size && is_function_pointer == other.is_function_pointer && is_integral == other.is_integral;
+    }
+  };
+
+  std::vector<MemberInfo> members;
+
+  bool operator==(const StructInfo& other) const {
+    return (const SimpleTypeInfo&)*this == (const SimpleTypeInfo&)other &&
+           std::equal(members.begin(), members.end(), other.members.begin(), other.members.end());
+  }
+};
+
+struct TypeInfo : std::variant<std::monostate, SimpleTypeInfo, StructInfo> {
+  using Parent = std::variant<std::monostate, SimpleTypeInfo, StructInfo>;
+
+  TypeInfo() = default;
+  TypeInfo(const SimpleTypeInfo& info)
+    : Parent(info) {}
+  TypeInfo(const StructInfo& info)
+    : Parent(info) {}
+
+  // Opaque declaration with no full definition.
+  // Pointers to these can still be passed along ABI boundaries assuming
+  // implementation details are only ever accessed on one side.
+  bool is_opaque() const {
+    return std::holds_alternative<std::monostate>(*this);
+  }
+
+  const StructInfo* get_if_struct() const {
+    return std::get_if<StructInfo>(this);
+  }
+
+  StructInfo* get_if_struct() {
+    return std::get_if<StructInfo>(this);
+  }
+
+  const SimpleTypeInfo* get_if_simple_or_struct() const {
+    auto as_struct = std::get_if<StructInfo>(this);
+    if (as_struct) {
+      return as_struct;
+    }
+    return std::get_if<SimpleTypeInfo>(this);
+  }
+};
+
+struct FuncPtrInfo {
+  std::array<uint8_t, 32> sha256;
+  std::string result;
+  std::vector<std::string> args;
+};
+
+struct ABI : std::unordered_map<std::string, TypeInfo> {
+  std::unordered_map<std::string, FuncPtrInfo> thunked_funcptrs;
+  int pointer_size; // in bytes
+};
 
 struct FunctionParams {
   std::vector<clang::QualType> param_types;
@@ -112,7 +193,8 @@ struct NamespaceInfo {
 
 class AnalysisAction : public clang::ASTFrontendAction {
 public:
-  AnalysisAction() {
+  AnalysisAction(const ABI& guest_abi)
+    : guest_abi(guest_abi) {
     decl_contexts.push_back(nullptr); // global namespace (replaced by getTranslationUnitDecl later)
   }
 
@@ -165,6 +247,8 @@ protected:
   RepackedType& LookupType(clang::ASTContext& context, const clang::Type* type) {
     return types.at(context.getCanonicalType(type));
   }
+
+  const ABI& guest_abi;
 };
 
 inline std::string get_type_name(const clang::ASTContext& context, const clang::Type* type) {
